Move std::hash & rlp encoding of Hash [skip ci]
